+++ test/TopologicalSortTest.cpp
@@ -96,7 +96,7 @@ TEST(TopologicalSortTest, test_3) {
 
   // check topological order of nodes
   std::unordered_map<CXXGraph::id_t, int> nodeToOrder;
-  for (int i = 0; i < res.nodesInTopoOrder.size(); ++i) {
+  for (size_t i = 0; i < res.nodesInTopoOrder.size(); ++i) {
     nodeToOrder[res.nodesInTopoOrder[i].getId()] = i;
   }
